# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tole Community Management API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Tole Community Management System",
      "url": "https://github.com/"
    },
    "license": {
      "name": "UNLICENSED"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "This server — all versioned endpoints live under the /api prefix."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the `token` returned by `POST /auth/login` into the Authorize dialog. It is sent as `Authorization: Bearer <token>`. Query-string tokens are rejected by the server."
      }
    },
...
    "parameters": {
...
      "ComplaintStatusQuery": {
        "name": "status",
        "in": "query",
        "required": false,
        "description": "Filter complaints by status.",
        "schema": {
          "$ref": "#/components/schemas/ComplaintStatus"
        }
      },
...
      "HouseIdQuery": {
        "name": "houseId",
        "in": "query",
        "required": false,
        "description": "Restrict results to one house. Residents are silently intersected with their own linked houses, so they can never widen their scope with this parameter.",
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-fA-F]{24}$",
          "example": "65f1a2b3c4d5e6f7a8b9c0d2"
        }
      },
      "SearchQuery": {
        "name": "search",
        "in": "query",
        "required": false,
        "description": "Case-insensitive substring match. The value is regex-escaped, so it is matched literally. Must be 100 characters or fewer — a longer value returns 400.",
        "schema": {
          "type": "string",
          "maxLength": 100,
          "example": "water"
        }
      },
      "SectionQuery": {
        "name": "section",
        "in": "query",
        "required": false,
        "description": "Filter by community section.",
        "schema": {
          "type": "string",
          "example": "Section 1"
        }
      },
...
      "CategoryQuery": {
        "name": "category",
        "in": "query",
        "required": false,
        "description": "Filter complaints by category.",
        "schema": {
          "$ref": "#/components/schemas/ComplaintCategory"
        }
      },
...
      "HistoryQuery": {
        "name": "history",
        "in": "query",
        "required": false,
        "description": "Set to the string `true` to include records belonging to archived houses. Any other value (including omitting the parameter) excludes them.",
        "schema": {
          "type": "string",
          "enum": [
            "true"
          ],
          "example": "true"
        }
      },
      "MineQuery": {
        "name": "mine",
        "in": "query",
        "required": false,
        "description": "Set to the string `true` to restrict complaints to those assigned to the calling staff member. Ignored for other roles.",
        "schema": {
          "type": "string",
          "enum": [
            "true"
          ],
          "example": "true"
        }
      },
      "SlaBreachedQuery": {
        "name": "slaBreached",
        "in": "query",
        "required": false,
        "description": "Set to `true` to return open complaints that have breached their configured SLA or have already been escalated.",
        "schema": {
          "type": "string",
          "enum": [
            "true"
          ],
          "example": "true"
        }
      },
...
    },
...
  },
  "paths": {
...
    "/complaints": {
      "get": {
        "operationId": "listComplaints",
        "summary": "List complaints",
        "description": "Returns complaints, newest first, with `submittedBy`, `assignedTo` and `resolvedBy` populated as `name`/`phone` (plus `specialization` for `assignedTo`).\n\nEach complaint includes `createdAt` (submitted), `startedAt` (when work began, if started), and `resolvedAt` (when resolved, if resolved).\n\nEach complaint includes a computed `slaTargetAt`. Open complaints past that target are marked `SLA Overdue`; escalated complaints carry `escalated: true` and `escalatedAt`. Pass `slaBreached=true` to filter the list to overdue or escalated open complaints.\n\n**Who sees what:**\n\n- **Residents** see only their own complaints.\n- **Staff** see all complaints; adding `mine=true` narrows the list to those assigned to them.\n- **Admins** see everything.\n\n`search` matches the title or description case-insensitively (regex-escaped, max 100 characters). By default, complaints belonging to archived houses are excluded — pass `history=true` to include them.\n\n**Pagination is opt-in** — omit `page` for the full list.",
        "tags": [
          "Complaints"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ComplaintStatusQuery"
          },
          {
            "$ref": "#/components/parameters/CategoryQuery"
          },
          {
            "$ref": "#/components/parameters/SectionQuery"
          },
          {
            "$ref": "#/components/parameters/HouseIdQuery"
          },
          {
            "$ref": "#/components/parameters/SearchQuery"
          },
          {
            "$ref": "#/components/parameters/MineQuery"
          },
          {
            "$ref": "#/components/parameters/SlaBreachedQuery"
          },
          {
            "$ref": "#/components/parameters/HistoryQuery"
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Matching complaints with pagination metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 20
                    },
                    "total": {
                      "type": "integer",
                      "example": 96
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "pages": {
                      "type": "integer",
                      "example": 5
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Complaint"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      },
      "post": {
        "operationId": "createComplaint",
        "summary": "File a complaint",
        "description": "Files a complaint, optionally with photo/PDF evidence.\n\n**Evidence:** send `multipart/form-data` and repeat the `attachments` field for each file — up to **5** files, each a JPEG/PNG/GIF image or PDF, 5 MB per file. Types are detected from the file's actual bytes, so renaming a `.exe` to `.jpg` will not help.\n\n**Priority:** residents may only file `low` or `medium`; `high`/`urgent` returns 403. Staff and admins may file any priority.\n\n**Section and house:** residents must have a house linked to their account (otherwise 400) and may only file against their own linked houses. `section` is normally auto-filled from that house.\n\n**What the response tells you.** Creating a complaint triggers three server-side analyses, and all three are returned so the UI can surface them:\n\n1. `autoDetected` — the `category` inferred from the text by keyword scoring, with a `confidence` value.\n2. `autoAssigned` — the staff member picked by specialisation match and current workload, or `null` if nobody matched. A `null` here means the complaint is still `pending` and waiting for manual assignment.\n3. `similarComplaints` — up to 3 previously resolved/closed complaints in the same category, ranked by TF-IDF cosine similarity, so the submitter can see how similar issues were handled. Auto-assigned complaints start as `inprogress` with `startedAt` set.",
        "tags": [
          "Complaints"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplaintRequest"
              }
            }
          },
          "description": "Text fields plus up to 5 optional attachments. Use this endpoint's \"Try it out\" form, not raw JSON."
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "The complaint was created, auto-categorised, auto-assigned, and matched against similar past complaints.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Complaint"
                    },
                    "autoDetected": {
                      "type": "object",
                      "properties": {
                        "category": {
                          "$ref": "#/components/schemas/ComplaintCategory"
                        },
                        "confidence": {
                          "type": "number",
                          "example": 0.82,
                          "description": "Keyword-scoring confidence between 0 and 1."
                        },
                        "section": {
                          "type": "string",
                          "example": "Section 1"
                        }
                      }
                    },
                    "autoAssigned": {
                      "nullable": true,
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/UserSummary"
                        }
                      ],
                      "description": "The staff member auto-assigned, or `null` when no specialisation matched. Expect `name`, `specialization` and `phone`."
                    },
                    "similarComplaints": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SimilarComplaint"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/UploadErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
    "/complaints/{id}/attachments/{filename}": {
      "get": {
        "operationId": "getComplaintAttachment",
        "summary": "Download a complaint attachment",
        "description": "Streams one attachment file. **Access is re-checked on every request** and is limited to the submitter, the assigned staff member, and admins — everyone else gets 403.\n\nThe `filename` must match a value stored in that complaint's `attachments` array exactly; the lookup is done against the parent record rather than the filesystem, and path traversal attempts are rejected.",
        "tags": [
          "Complaints"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          },
          {
            "$ref": "#/components/parameters/FilenamePathParam"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ImageOrPdfResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
    "/complaints/{id}": {
      "put": {
        "operationId": "updateComplaint",
        "summary": "Update a complaint",
        "description": "Change a complaint's status, assignment or resolution text.\n\n**Status state machine**\n\n| From | To | Who |\n| --- | --- | --- |\n| `pending` | `inprogress` | assigned staff, or admin |\n| `inprogress` | `resolved` | assigned staff, or admin — `resolution` is **required** |\n| `resolved` | `pending` | the original submitter (a reopen), or admin/staff |\n| `resolved` | `closed` | admin/staff |\n| `closed` | *(nothing)* | **locked forever** |\n\nAny other move returns **400**.\n\nAdditional rules:\n- Staff may only update complaints **assigned to them**; everyone else gets 403. Admins bypass this.\n- Residents may only act on their **own** complaints, and only to reopen them.\n- `resolution` must be 10-500 characters when resolving.\n- `assignedTo` is silently ignored for callers who are not admin or staff.\n- Once `reopenCount` reaches 2 the complaint is **forced** to `closed`, whatever was requested.",
        "tags": [
          "Complaints"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPathParam"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComplaintRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The updated complaint.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Complaint"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/ServerErrorResponse"
          }
        }
      }
    },
...
  }
}